SchemaRegistry: Defensively copy registered ref
Make a copy of a ref schema before returning it to the caller after
registering a new type to prevent the caller from corrupting the
registry entry.

This is more likely now that $ref is valid alongside other properties.

We already did a copy when a ref was retrieved from the registry, but
not when it was first registered.